+++ my.kv
@@ -1,9 +1,9 @@
 WindowManager:
     MainWindow:
-    SecondWindow:
-    ThirdWindow:
-    FourthWindow:
-
+    PlayChessWindow:
+    WatchChessWindow:
+    DemoWindow:
+    ShutDown:
 <MainWindow>:
     name: "main"
     FloatLayout:
@@ -11,88 +11,92 @@ WindowManager:
             Rectangle:
                 pos: self.pos
                 size: self.size
-                source: 'background.png'
+                source: 'GUIPics/background.png'
         
         ImageButton:  
-            source:'PlayChess.png'
+            source:'GUIPics/PlayChess.png'
             size_hint: 0.15, 0.15  
             pos_hint: {"x": 0.25, 'y':0.55}
             on_release:
-                app.root.current = "second"
+                app.root.current = "playchess"
                 root.manager.transition.direction = "left"
-
         ImageButton:  
-            source:'Power.png' 
+            source:'GUIPics/Power.png' 
             size_hint: 0.15, 0.15 
             pos_hint: {"x": 0.58, 'y':0.55}
             on_release:
-                app.root.current = "second"
+                app.root.current = "shutdown"
                 root.manager.transition.direction = "left"            
-
         ImageButton:  
-            source:'WatchChess.png'  
+            source:'GUIPics/WatchChess.png'  
             pos_hint: {"x": 0.58, 'y':0.25}
             size_hint: 0.15, 0.15
             on_release:
-                app.root.current = "third"
+                app.root.current = "watchchess"
                 root.manager.transition.direction = "left"
         
         ImageButton:  
-            source:'DemoMode.png'  
+            source:'GUIPics/DemoMode.png'  
             pos_hint: {"x": 0.25, 'y':0.25}
             size_hint: 0.15, 0.15
             on_release:
-                app.root.current = "fourth"
+                app.root.current = "demoscreen"
                 root.manager.transition.direction = "left"
-
-
-<SecondWindow>:
-    name: "second"
+<PlayChessWindow>:
+    name: "playchess"
     FloatLayout:
         canvas.before:
             Rectangle:
                 pos: self.pos
                 size: self.size
-                source: 'background.png'
-
+                source: 'GUIPics/background.png'
         ImageButton:
-            source:'BackArrow.png'
+            source:'GUIPics/BackArrow.png'
             pos_hint: {"x": 0.90, 'y':0.87}
             size_hint: 0.10, 0.10
             on_release:
                 app.root.current = "main"
                 root.manager.transition.direction = "right"
-
-<ThirdWindow>:
-    name: "third"
-
+<WatchChessWindow>:
+    name: "watchchess"
     FloatLayout:
         canvas.before:
             Rectangle:
                 pos: self.pos
                 size: self.size
-                source: 'background.png'
-
+                source: 'GUIPics/background.png'
         ImageButton:
-            source:'BackArrow.png'
+            source:'GUIPics/BackArrow.png'
             pos_hint: {"x": 0.90, 'y':0.87}
             size_hint: 0.10, 0.10
             on_release:
                 app.root.current = "main"
                 root.manager.transition.direction = "right"
-
-<FourthWindow>:
-    name: "fourth"
-
+<DemoWindow>:
+    name: "demoscreen"
     FloatLayout:
         canvas.before:
             Rectangle:
                 pos: self.pos
                 size: self.size
-                source: 'background.png'
-
+                source: 'GUIPics/background.png'
         ImageButton:
-            source:'BackArrow.png'
+            source:'GUIPics/BackArrow.png'
+            pos_hint: {"x": 0.90, 'y':0.87}
+            size_hint: 0.10, 0.10
+            on_release:
+                app.root.current = "main"
+                root.manager.transition.direction = "right"
+<ShutDown>:
+    name: "shutdown"
+    FloatLayout:
+        canvas.before:
+            Rectangle:
+                pos: self.pos
+                size: self.size
+                source: 'GUIPics/background.png'
+        ImageButton:
+            source:'GUIPics/BackArrow.png'
             pos_hint: {"x": 0.90, 'y':0.87}
             size_hint: 0.10, 0.10
             on_release:
